RavenDB-20366 Separate global and local databases slice
